refactor: adopt latest protobufs (#34)
Because

- the latest protobuf requires the adoption for pipeline-backend

This commit

- update with the latest protobuf which simplify output and update patch
- close #32

+++ pkg/model/pipeline.go
@@ -36,7 +36,7 @@ type Pipeline struct {
 type ListPipelineQuery struct {
 	WithRecipe bool
 	Namespace  string
-	PageSize   int32
+	PageSize   uint64
 	Cursor     uint64
 }
